+++ tests/testthat/test-density_weights.R
@@ -22,7 +22,7 @@ dens_lrn <- with(
   haldensify(
     A = A, W = W,
     wts = Weights,
-    n_bins = c(5, 10, 15),
+    n_bins = c(3, 5, 10),
     lambda_seq = exp(seq(-1, -13, length = 200))
   )
 )