+++ R/haldensify.R
@@ -187,17 +187,17 @@ cv_haldensify <- function(fold,
 #' @examples
 #' # simulate data: W ~ U[-4, 4] and A|W ~ N(mu = W, sd = 0.5)
 #' set.seed(429153)
-#' n_train <- 50
+#' n_train <- 100
 #' w <- runif(n_train, -4, 4)
 #' a <- rnorm(n_train, w, 0.5)
 #' # learn relationship A|W using HAL-based density estimation procedure
 #' haldensify_fit <- haldensify(
 #'   A = a, W = w, n_bins = c(3, 5),
-#'   lambda_seq = exp(seq(-1, -5, length = 50)),
-#'   reduce_basis = 0.3, max_degree = 3
+#'   lambda_seq = exp(seq(-1, -5, length = 500)),
+#'   # the following arguments are passed to hal9001::fit_hal()
+#'   max_degree = 3, reduce_basis = 0.2, smoothness_orders = 0
 #' )
-haldensify <- function(A,
-                       W,
+haldensify <- function(A, W,
                        wts = rep(1, length(A)),
                        grid_type = "equal_range",
                        n_bins = c(3, 5, 10),
@@ -205,12 +205,13 @@ haldensify <- function(A,
                        lambda_seq = exp(seq(-1, -13, length = 1000)),
                        hal_basis_list = NULL,
                        ...) {
-  # capture dots
+  # capture dot arguments
   dots <- list(...)
 
   # if W is set to NULL, create a constant conditioning set
+  # NOTE: this essentially recovers the marginal density of A
   if (is.null(W)) {
-    W <- rep(0, length(A))
+    W <- rep(1, length(A))
   }
 
   # run CV-HAL for all combinations of n_bins and grid_type